fix(ci): python-coverage-shards — split oversized gc_worker_delegation test module under 1500-line guardrail
test_gc_worker_delegation.py grew to 1529 lines, tripping
test_first_party_code_files_stay_under_line_limit. Move the headline
candidate/preserved-list folding tests into a new
test_gc_worker_delegation_headline_lists.py module; both files are now
well under MAX_FIRST_PARTY_FILE_LINES.
